@@ -76,6 +76,6 @@
     "tailwindcss": "^3.3.5"
   },
   "engines": {
-    "node": ">=14.x"
+    "node": ">=20.x"
   }
 }
